feat: adding hamburger menu switch
