+++ src/components/MobileMenu/MobileMenu.scss
@@ -1,0 +1,10 @@
+.menu-switch {
+	filter: invert(100%);
+	padding: 12px;
+	margin: 12px;
+	border-radius: 50%;
+	border: 2px solid black;
+	background-color: white;
+  width: 24px;
+  height: 24px;
+}
