--- src/index.css
@@ -1,93 +1,0 @@
-/* :root {
-	font-family: Montserrat, system-ui, Avenir, Helvetica, Arial, sans-serif;
-	line-height: 1.5;
-	font-weight: 400;
-
-	color-scheme: light dark;
-	color: rgba(255, 255, 255, 0.87);
-	background-color: #242424;
-
-	font-synthesis: none;
-	text-rendering: optimizeLegibility;
-	-webkit-font-smoothing: antialiased;
-	-moz-osx-font-smoothing: grayscale;
-}
-
-body {
-	margin: 0;
-	place-items: center;
-	min-width: 320px;
-	min-height: 100vh;
-	color-scheme: light dark;
-}
-
-h1 {
-	font-size: 3.2em;
-	line-height: 1.1;
-}
-
-button {
-	border-radius: 8px;
-	border: 1px solid transparent;
-	padding: 0.6em 1.2em;
-	font-size: 1em;
-	font-weight: 500;
-	font-family: inherit;
-	background-color: #1a1a1a;
-	cursor: pointer;
-	transition: border-color 0.25s;
-}
-button:hover {
-	border-color: #646cff;
-}
-button:focus,
-button:focus-visible {
-	outline: 4px auto -webkit-focus-ring-color;
-}
-
-@media (prefers-color-scheme: light) {
-	:root {
-		color: #213547;
-		background-color: #ffffff;
-	}
-	a:hover {
-		color: #747bff;
-	}
-	button {
-		background-color: #f9f9f9;
-	}
-}
- */
-body {
-	margin: 0;
-	place-items: center;
-	min-width: 320px;
-	min-height: 100vh;
-	color-scheme: light dark;
-}
-
-.body-theme__dark {
-	background-color: #242424;
-	color: rgba(255, 255, 255, 0.87);
-}
-
-.body-theme__light {
-	background-color: #ffffff;
-	color: #213547;
-}
-
-@media (prefers-color-scheme: dark) {
-	body {
-		/* Estilos para el esquema de color oscuro */
-		background-color: black;
-		color: white;
-	}
-}
-
-@media (prefers-color-scheme: light) {
-	body {
-		/* Estilos para el esquema de color oscuro */
-		background-color: white;
-		color: black;
-	}
-}
