+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "./index.css";
+import "./index.scss";
 import App from "./components/App/App.jsx";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
