+++ src/index.scss
@@ -1,0 +1,53 @@
+$color-darker: #071220;
+$color-accent1: #B32570;
+$color-accent2: #FECD1A;
+$color-background: #12253D;
+$color-text: #fff;
+
+
+:root {
+	font-family: Montserrat, Avenir, Helvetica, Arial, sans-serif;
+	line-height: 1.5;
+	font-weight: 400;
+
+	color-scheme: light dark;
+
+	font-synthesis: none;
+	text-rendering: optimizeLegibility;
+	-webkit-font-smoothing: antialiased;
+	-moz-osx-font-smoothing: grayscale;
+}
+
+body {
+	margin: 0;
+	place-items: center;
+	min-width: 320px;
+	min-height: 100vh;
+	color-scheme: light dark;
+}
+
+.body-theme__dark {
+	background-color: $color-background;
+	color: $color-text;
+}
+
+.body-theme__light {
+	background-color: $color-text;
+	color: $color-background;
+}
+
+@media (prefers-color-scheme: dark) {
+	body {
+		/* Estilos para el esquema de color oscuro */
+		background-color: $color-background;
+		color: $color-text;
+	}
+}
+
+@media (prefers-color-scheme: light) {
+	body {
+		/* Estilos para el esquema de color oscuro */
+		background-color: $color-text;
+		color: $color-background;
+	}
+}
